Сделаны обязательный задания дз 7
