+++ src/components/MessageList/MessageList.jsx
@@ -1,5 +1,4 @@
 import {Message} from "../Message/Message";
-import {AUTHORS} from "../../utils/constants";
 import './MessageList.styles.css';
 import {Container} from "@mui/material";
 import PropTypes from "prop-types";
@@ -11,8 +10,7 @@ export const MessageList = ({messages}) => {
     return (
         <Container maxWidth={"sm"}>
             <div className="messageArea">
-                {messages.map((message) => (<Message key={message.id} msgText={message.text} author={message.author}
-                                                     robotName={AUTHORS.robotName}/>))}
+                {messages.map((message) => (<Message key={message.id} msgText={message.text} author={message.author}/>))}
             </div>
         </Container>)
 }
