+++ src/index.js
@@ -3,13 +3,16 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import {store} from "./store";
+import {persistor, store} from "./store";
 import {Provider} from "react-redux";
+import {PersistGate} from "redux-persist/integration/react";
 
 ReactDOM.render(
     <React.StrictMode>
         <Provider store={store}>
-            <App/>
+           <PersistGate persistor={persistor}>
+               <App/>
+           </PersistGate>
         </Provider>
     </React.StrictMode>,
     document.getElementById('root')
