+++ src/screens/Chat/Chat.jsx
@@ -1,27 +1,25 @@
-import {useEffect, useMemo, useRef} from "react";
+import {useMemo, useRef} from "react";
 import {Navigate, useOutletContext, useParams} from "react-router-dom";
 
 import './Chat.css'
-import {AUTHORS} from "../../utils/constants";
 import {MessageList} from "../../components/MessageList/MessageList";
 import {Form} from "../../components/Form/Form";
 import {Button} from "@mui/material";
 import {shallowEqual, useDispatch, useSelector} from "react-redux";
-import {addMessage} from "../../store/messages/actions";
+import {addMessageWithReply} from "../../store/messages/actions";
 import {selectMessagesById} from "../../store/messages/selectors";
 import {selectName} from "../../store/profile/selectors";
 
 export const Chat = () => {
     const {id} = useParams();
     const handleDeleteChatClick = useOutletContext();
-    const getMessages = useMemo(()=> selectMessagesById(id), [id]);
+    const getMessages = useMemo(() => selectMessagesById(id), [id]);
     const messageList = useSelector(getMessages, shallowEqual);
     const author = useSelector(selectName);
     const dispatch = useDispatch();
-    const timeout = useRef();
 
     const addMsg = (Msg) => {
-        dispatch(addMessage(id, Msg));
+        dispatch(addMessageWithReply(id, Msg));
     }
 
     const sendMsg = (text) => {
@@ -31,26 +29,6 @@ export const Chat = () => {
             id: `msg-${Date.now()}`,
         })
     }
-
-
-    useEffect(() => {
-        const lastMsg = messageList?.[messageList?.length - 1];
-        if (lastMsg?.author === author) { //optional chaining
-            timeout.current = setTimeout(() => {
-                addMsg({
-                    author: AUTHORS.robotName,
-                    text: "Ваше сообщение отправлено",
-                    id: `msg-${Date.now()}`,
-                })
-            }, 1500);
-        }
-
-
-        return () => {
-            clearTimeout(timeout.current);
-        };
-
-    }, [messageList])
 
     if (!messageList) {
         return <Navigate to={"/chat"} replace/>
