+++ src/store/index.jsx
@@ -1,12 +1,28 @@
-import {combineReducers, createStore} from "redux";
+import {applyMiddleware, combineReducers, compose, createStore} from "redux";
 import {profileReducer} from "./profile/reducer";
 import {chatsReducer} from "./chats/reducer";
 import {messagesReducer} from "./messages/reducer";
+import thunk from "redux-thunk";
+import storage from "redux-persist/lib/storage";
+import {persistReducer, persistStore} from "redux-persist";
 
+const persistConfig =  {
+    key: "Messenger",
+    storage
+}
 const rootReducer = combineReducers({
-    profile:profileReducer,
-    chats:chatsReducer,
-    messages:messagesReducer,
+    profile: profileReducer,
+    chats: chatsReducer,
+    messages: messagesReducer,
 })
 
-export const store = createStore(rootReducer, window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__())
+const persistedReducer = persistReducer(persistConfig,rootReducer);
+
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+
+export const store = createStore(
+    persistedReducer,
+    composeEnhancers(applyMiddleware(thunk))
+)
+
+export const persistor = persistStore(store);
